feat(cq): moved func to domain
refs. metriport/metriport-internal#1352
